ada_server: refactor sender+receiver creation & start
No functional change should happen here except of logging messages.
The Ada Server logged events for sender to both sender and receiver
logger (the same for receiver). This is probably a relict of a
debugging process. As we separate creation of sender and receiver
here, no such logging happens. Anyway, the two threads should not
depend on each other.