ui_server: introduce debug utilities
Provide several macros and the class Trace which are helpful for
debugging and tracing of the code. The Trace class is not an API
and should not be used directly. Instead, there are 6 useful macros:

* LOGGER_CLASS(instance)
* LOGGER_FUNC(func)
* TRACE_METHOD()
* TRACE_FUNC()
* _TRACE_METHOD(logger)
* _TRACE_FUNC(logger)

providing an API for debugging and tracing. The LOGGER_* macros
are useful for creating appropriate loggers for a certain instance
or function.

Reviewed-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>