ui_server: include mongoose libarary
The mongoose library is under the GNU/GPL license and thus it MUST
NOT be merged upstream in this form as it would break the current
BSD license of the project. Mongoose is used for testing of the
HTTP server and it will be replaced by some license-compatible
solution (civetweb).

Reviewed-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>