ui_server: new generic class Entity
Entity is the base class for inter-layer data objects. It is not
an interface. Its purpose is to provide serialization and
deserialization capabilities and to simplify implementation
of the data objects. Each Entity as a generic EntityCollection
(can be overriden in its childs).

The Entity is not a subject of the service layer that should
never directly access its methods.

Reviewed-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>