SociManager: new class
SociManager manages a pool of connections to a database. All soci
backends can be used if configured. The SociManager also optionally
executes an initialization SQL script. This is useful for deployment
to a new server, to execute update commands or to setup a testing
environment.

The SociManagerTest has accidently been included for building of
tests in 01275862938e.

Fixes: 01275862938e ("Base64: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>