Credentials: improve structure to support token-based auth
The initial Credentials implementation assumed a different
login scheme. In this approach, we either receive the Authorization
header containing an access token or we require credentials.

The credentials are expected as a pair (provider, token).
Thus, we introduce class TokenCredentials as a specialization
to support this.

Related code is updated accordingly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>