SecureRandomProvider: fix random bytes generation
The randomString method should have provided a random string of
a given length. However, the generated string has contained any
bytes (i.e. also non-printable ones). The method is renamed
to randomBytes and the conversion to string is now up to the
caller.

Update SessionManager to encode the random bytes by Base64
algorithm. The session ID is now a string of printable characters.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>