Device: define database schema and contents
The device is different from the orignal approach. The init attribute
is replaced by active_since that can be NULL or the time of the
activation by a user. We also record the first occurence of the
device by a certain gateway and the last time when it was available.

The type of device is just a number for now. Later, we should
introduce some type model into the system.

Signed-off-by: Petr Stehlík <xstehl14@stud.fit.vutbr.cz>
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>