t: Rest: change call API to avoid leaks
While performing a high number of request via the Rest class,
python starts receiving timeouts from poll system call:

  connect(27, {sa_family=AF_INET, sin_port=htons(10000),
		sin_addr=inet_addr("127.0.0.1")}, 16)
		= -1 EINPROGRESS (Operation now in progress)
  poll([{fd=27, events=POLLOUT|POLLERR}], 1, 1000)
		= 1 ([{fd=27, revents=POLLOUT}])
  getsockopt(27, SOL_SOCKET, SO_ERROR, [0], [4]) = 0
  setsockopt(27, SOL_TCP, TCP_NODELAY, [1], 4) = 0
  poll([{fd=27, events=POLLOUT}], 1, 1000) = 1 ([{fd=27, revents=POLLOUT}])
  sendto(27, "DELETE /auth HTTP/1.1\r\nHost: loc"..., 187, 0, NULL, 0) = 187
  poll([{fd=27, events=POLLIN}], 1, 1000) = 0 (Timeout)

Probably, the reason is the __del__ function which acts as a
destructor, however, it is not very safe in the context of
Python. See:

  http://www.algorithm.co.il/blogs/programming/python-gotchas-1-
  	__del__-is-not-the-opposite-of-__init__/

To solve this situation (several tests fail due to the timeout),
the API of the class Rest is slightly modified. Now, the __call__()
returns a pair: (response, content) and the connection is closed
immediately.

Fixes: c871ee1b9470 ("t: add library rest")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>