DependencyInjector: allow creation of disowned instances
Some instances may be deallocated by other facility then
the DependencyInjector itself. This is sometimes true for
certain 3rd party libraries. We can either wrap such object
in another instance and avoid implementing its destructor
or ask DependencyInjector to disown such instance.

A disowned instance is not managed by the DependencyInjector
anymore and thus must be deleted manually. Also, it is not
a singleton from the point of view of the DependencyInjector.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>