Pass session as UserData
This step requires many changes as the UserData is passed through
many templates. Thus, the templating has been simplified or removed.
PocoServer and MongooseServer are now both non-templates which
greatly simplifies the situation.

The Route is untouched and still accepts anything as UserData.
However, the request handling does not use any UserData anymore.
Instead, it obtains the ExpirableSession::Ptr from SessionVerifier
and feeds the Route with such instance.

The original intention to have request handlers implemented as
static functions has been dropped so there is no need to pass
any object with services any more. Each handler is now a proper
class and has its own dependencies given by dependency injection.
So, the UserData is now free to be utilized for passing the session.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>