LocationHandler: apply access policy
Now, when a non-existing location is requested, we can either say
"not found" or "forbidden" regardless if it exists or not. We
choose to send forbidden to avoid leaking of information.

When accessing /place/P0/locations/L0 but L0 is in place P1
and we have access to both P0 and P1 then we expect 404.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>