PlaceHandler: implement delete properly
When deleting a place, we have to check there is no other
role then those of the deleting user. Those roles must be
deleted and then it is possible to drop the place.

Now, the GET on a deleted place should expect 403 "forbidden"
as we do not want to disclose whether such place exists
or it is just unavailable to the user.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>