SessionManager: open returns ExpirableSession::Ptr
Make the API more high-level. When we open a new session, it
is OK to have the ExpirableSession instance available for
direct access. Otherwise, it would cost a lookup call.

The implementation, however, calls the cache->get(id) which
performs that unnecessary lookup. The reason is currently
the fact that the AbstractCache has no return value with
the copied value inserted into it. This situation is not
ideal but we can live with this for some time because the
API is free from this issue.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>