Deserializer: new template interface for deserialzation
The update procedure cannot be easily implemented in the service
layer because we need to first fetch an object from Dao, then
apply the deserialized data and finally call an update.

To make this possible, introduce an Deserializer abstraction that
handles the deserizalization later when it is needed deeper
in the service layer.

This also allows to move more code into the service layer.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>