PocoDaoManager: new class
PocoDaoManager represents an access point to a database.
The Dao implementations would obtain the SessionPool from
the PocoDaoManager to perform SQL queries. The PocoDaoManager
can initialize database by a startup SQL script.

The initial script is considered to be a template. So all occurences
of ${...} can be replaced before the actual execution. This might
help to provide a backend-independent SQL schema.

We occasionally log the session pool usage and provide information
about transaction properties of the initial session on startup.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>