UIMockInit: prepare initialized devices for testing
The humidity device is prepared for availability tests. After
the server is started, the humidity would have the lastSeen
set to the current time.

Other devices would be always unavailable due to fixed lastSeen
timestamps.

All types are zero as there is no clear semantics for this value
at the moment.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>