PlaceService: extract method hasUsersExcept
Move this logic out of the service layer. First, it uses IDs
of objects (this must not happen on service layer). Second, it
prevents the removal of the Identity-User relationship.

The solution is not optimal and should be improved in the future.
Except of a possible race conditions, the PlaceService still
does too low-level job when removing roles.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>