AuthService: fix login by different provider
When logging in by a different provider, we reuse the User
instance that has already been created instead of creating
a new one.

Note that there is a possible situation with many VerifiedIdentity
instances pointing to Identity and one of the VerifiedIdentity
instances is assigned to a different User. Such situation is
WRONG and there should be an integrity check on the Dao level.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>