GatewayService: always create an implicit place on register
The registerGateway always creates an implicit Place for the
gateway. The original behaviour was trying to reuse an existing
one. But there is no way how to decide the right place for this
purpose and such Gateway might end up in a Place of somebody
else. By having the Place-Gateway 1:1 correspondence, we also
preserve the old xmlui behaviour as there is no Place abstration
visible, however, access control is done per Gateway.

Moreover, the registerGateway is now simplier.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>