DeviceService: implement fetchMany
To satisfy the xml API ns="devices" type="get" request which
can ask selectively for devices, implement the fetchMany
service method. For each device, it loads that devices if
possible. Inaccessible devices are skipped silently.

The implementation is done on the service layer, however, a more
optimal way is to move to the Dao layer. This is to be done later.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>