PrivateKeyPassphraseProvider: new class
The standard Poco way of dealing with private key passphrases
does not allow to provide different passphrases for different
keys. The reason is probably the OpenSSL API which seems to
does not provide enough information itself.

We solve this by locking all context initializations (required
to share the appropriate mutex) and so we know which context
is initialized. This way can provide the right password to
the right private key.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>