OAuth2AuthProvider: configure SSL from SSLClient
There is no need to have multiple SSL contexts when asking
the same servers for authorization of a user. Also note that
doing that is not thread-safe because the InvalidCertificateHandler
registers itself globally to SSLManager. Moreover, using the
SSLManager::initializeClient is not thread-safe because it sets
a global context and handlers.

Note that we need to ensure that the GoogleAuthProvider accepts
the Google certificates. To enable this, the google.ssl.authority
should point to a CA storage.

The default redirect_uri is set to http://localhost as it seems
to be sufficient for now.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>