GoogleAuthProvider: drop obsolete TODO comment
According to [1], it is possible to verify the validity of
an id_token. However, the use case is different from the
situation we have. The purpose is to verify the id_token
that a client sends to the server is valid. In the case
of our server, the id_token never leaves the server so
there is no need to verify it.

The id_token is gather via the secure HTTPS channel and
the it should be verified that it has a valid certificate.
If this condition is matched then we do not have any reason
to verify id_token itself.

[1] https://developers.google.com/identity/sign-in/web/backend-auth

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>