conf: GoogleAuthProvider: CA by mozilla 2016-11-02
To make sure we have the CA to verify Google, include the CA
certificates extracted from Mozilla [1]. It simplifies testing.
Deploying of this CA list is not a subject of this commit.

Set this CA list in testing configurations and make it default
for GOOGLE_CA_LOCATION in GoogleAuthProviderTest.

The CA list is distributed under the MPL 2.0 and it should be
compatible with the BSD license of this project.

[1] https://curl.haxx.se/docs/caextract.html

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>