NotificationService: convert to NotificationCenter
The NotificationService is not polymorphic anymore. It now serves
as a wrapper around NotificationCenter. The FakeNotificationService
was dropped and replaced by the ConsoleObserver.

The AuthServiceTest has been reworked to use NotificationObserver.
Note that the NotificationCenter makes a clone of each observer
and so the check for lastIdentity is done in quite a different way.
We store just a pointer to m_lastIdentity in the observer that is
copied even when cloning the instance.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>