LegacyGateway: implement serialization
Implement serialization of LegacyGateway as required by
the XML-UI API. The status of gateways is fixed for now
and will be implemented later. Similar for version and
timezone. The current (legacy) server does not work
correctly with those fields are some parts aren't
implemented well. Thus, we wait for future decisions
to implement those features.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>