schema: describe queries in XML file
All or most of SQL queries can be expressed in a XML file
together with documentation and tests. This simplifies SQL
query testing and management. The Dao layer can load the
queries from here on start-up.

The XML format is beneficial when considering multiple target
SQL engines with different syntax requirements.

A first example query is included.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>