schema: export queries into INI file
The SQL queries can be loaded from a file on application startup.
The provided XSL transformation produces an INI file with a set
of key = value pairs. The keys are known to the application code.

Execute as:

  $ xsltproc --stringparam engine sqlite \
	  ../tools/xsl/query.export.xsl home8-query.xml
  $ ls
  home8-sqlite-queries.ini

The INI file is generated automatically by the cmake infrastructure
so no manual intervention is needed normally.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>