query: support postgre in queries tests
Generate CSV file with expected data as exported by PostgreSQL.
Expects running the psql with -q option, eg.:

  $ cat home8-postgre-test.sql | psql -q testing-db

The -q avoids outputs like:

  INSERT 0 1
  DELETE 1
  ...

Those outputs is difficult to generate into the expect CSV file
and anyway it would make the CSV invalid.

The CSV headers are however avoided because then the psql generates
the headers everytime, even, when no output is expected. This is
not easy to solve with the current query.xml format. We do not have
the column names to put there in the source XML file.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>