query: test places.create as failing INSERT
Testing of insert queries is handled by triggers. If an insert
statement fails and the fail is expected (correct) now trigger
is called for the table. Thus, as we expect the insert to fail,
we prepare a trigger that generates an error message into a
special testing-only table with error messages and call a select
on this table. If it is empty, there is no error (the insert
failed as expected). If there is a message, it describes what
was wrong during this test.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>