SSLFacility: provide property disabledProtocols
The Poco library (since 1.7.0) supports to disable certain protocol
implementations as described in the following commit:

   e9b2b3bcdd7e ("NetSSL: add support for disabling certain protocols")

Implement API to utilize this feature.

The SSLv2 and SSLv3 are disabled as default. This can be overrided by
calling setDisabledProtocols or setting property disabledProtocols via
dependency injection. E.g.

   <instance ... class="BeeeOn::SSLClient">
	<set name="disabledProtocols" text="sslv2,sslv3,tlsv1" />
   </instance>

The SSLFacility logs a warning when the disabledProtocols feature is
unavailable due to the old Poco library.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>