schema: define views roles_in_gateway and legacy_gateways
To support the legacy XML API properly, we have to provide
more data then just the stored entries in the gateways table.
Some fields must be computed from the available relations.

First, we need to know all users that can access a gateway.
This is very easy to compute by a single SELECT statement.

Second, we need to know all devices active in a gateway.
This is also a trivial SELECT.

Third, the ID of the gateway's owner. An owner is the oldest
known admin for the gateway. So if Bob registers a gateway G,
he is the owner. If Steve is assigned an admin role for the
gateway G he does not become the owner. Bob is still the only
owner of the gateway G. However, when the Bob's role is deleted
or changed to a lower access level then Steve becomes the owner.

Finally, we need the access level of the user who is asking
the legacy gateway information. This is not covered here.
A query must be built for this purpose. The roles_in_gateway
view can simplify that query.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>