GatewayXmlHandler: implement handleGet via LegacyGateway
Squash of several closely related changes:

 * query: implement legacy_gateways.fetch.by.id
 * PocoSQLGatewayDao: implement fetch LegacyGateway
 * GatewayService: implement fetch of LegacyGateway

The query implements fetching of legacy gateways by ID. The
XML-UI API also requires to find out the access level to the
gateway by the user who is asking it.

The final handleGet implementations just exchanges using
of Gateway for LegacyGateway. The method overloading and
polymophism handle the rest.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>