schema: implement handling of tag <view />
The database schema can contain views. A view is usually
created by SQL syntax

   CREATE VIEW [name] AS SELECT [what] AS [as] FROM [base]
      JOIN [remote] ON [join-on-base] = [join-on-remote] ...;

This kind of view can be described by the construct:

   <view name="[name]" base="[base]">
      <map ref="[what0]" as="[as0]" />
      <map ref="[what1]" as="[as1]" />
      ...
      <join remote="[remote]" on-base="[join-on-base]" on-remote="[join-on-remote]">
          <map ref="[remote-what0]" as="[as(n+1)]" />
          <map ref="[remote-what1]" as="[as(n+2)]" />
	  ...
      </join>
      ...
      <sub-query as="[as(n+m+1)]">
          <sql>[what(n+m+1)]</sql>
      </sub-query>
   </view>

Thus, the [what] AS [as] can be defined by

* <map ref="[what]" as="[as]" />
* <join ...><map ref="[remote-what]" as="[as]" /></join>
* <sub-query as="[as]"><sql>[what]</sql></sub-query>

and the JOIN ON logic is expressed by the

    <join on-base="[left]" on-remote="[right]" />

with operation "equals" (=).

Multiple JOINs would be generated if requested in the input.
The JOIN always works with a base table. The default base
table is expressed by <view base="[base1]"> however, it can
be overriden by <join with-base="[base2]"> where the [base2]
is another joined table.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>