schema: generalize drop-tables algorithm
The drop-tables algorithm has been generalized to work with
views as well. It is now called cascade-top. The dependencies
are determined by template matching in mode cascade-dependencies.

The output of the algorithm is a node-set in format:

  <entry type="{type}">{name}</entry>

where {type} is "table" or "view" and {name} is its name.

The resulted node-set is converted to DROP sequence in SQL.
The sequence, however, must be sorted in reverse order for
to get the proper working DROP sequence.

The sub-query tags are processed by extract-table-and-view-names
template. Thus, dependencies in the custom SQL commands should
be detected as well.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>