query: reimplement test generation to be extendable
The current query.test.xsl assumes that the target engine
is selected by param $engine from command line. However,
this approach is not very suitable for XSLT because it cannot
dynamically handle the parameter very efficiently.

For every engine, there must be an entry file that overrides
the necessary templates to customize the test generation as
necessary. This is the same situation as would happen when
extending to more engines. Thus, in fact the amount of files
will not be higher.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>