schema: visualize view using dot
Views are drawn with a background color. All joins are drawn as an
intermediate box describing the ON condition. The sub-query fields
are connected by an intermediate ellipse named 'sub-query'.

The intermediate boxes and ellipses are useful to reduce the complex
noise made of indistinguishable arrows. All objects are further from
each other.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>