schema: find tables and views in text
Sometimes we need to determine whether a custom SQL code
references an existing table or view. Provide the template
extract-table-and-view-names for this purpose. It calls
apply-templates in mode extracted-from-text for each @name
node of a table or view that is included in the given text.

Also, it can preserve the context of the caller to make some
other computation possible.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>