query: mark gateways.fetch.by.place_id as obsolete
The query is to be removed when the application code stops
using it. It seems to be redundant.

The original sense of the query was to protected from malicious
request where the user asks for a gateway in a different place.
This case would be fixed differently. The gateway ID is always
unique and permissions are checked by another query.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>