schema: define table gateways_status
The table stores status of each gateway in time. It is assumed
that it won't hold many data as the gateway status is not being
changed very often.

It is thus possible to detect unexpected changing of IP addresses,
time when a gateway has been upgraded to a new version and more
when implemented.

It is however desirable to limit the size of the table by dropping
old records or archiving them out of the primary storage.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>