Gateway: support properties last_changed, version and ip
The version and ip are needed for XML-UI API.
The REST API reports them as well.

The values can be invalid. An invalid IP is reported as
0.0.0.0 and an invalid version as an empty string.

The check for IP are added to XML-UI tests as there was
no such attribute before.

Checks for all the fields are added REST API tests.

There is currently no way how to insert a gateway status
into the database from application. Thus, all expected
test data are the empty ones.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>