query: define type="uuid" for expected values
The isql prints uuid values in upper case. As we are unable
to detect value types (they do not depend on the input values
and we do not know which tables they are coming from), it is
necessary to mark every uuid field explicitly.

The uuid values are converted to upper case automatically for
the file with expected values.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>