query: change testing data to avoid double precision issues
The isql (but neigher of psql, sqlite3) does not print some double
values precisely. Change the values in tests to workaround this issue.
All double values that could be represented unprecisely are changed
to very similar ones with the decimal part set to something that can
be easily represented.

Note that we cannot use decimal part '.0' because other engines might
omit this when (and isql does) printing them. This would introduce
other similar diff issues.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>