PocoSQLLocationDao: implement method parseIfIDNotNull
The method is useful when a query returns location in
the result set while location is not the primary object
we are asking for.

Extracting of column name is left optional. However,
in PocoSQLPlaceDao, the name is not optional and requires
such queries to join the places table. It is not clear
whether we need this. Very probably, the way location
implements parseIfIDNotNull is better.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>