SSLFacility: preset m_disabledProtocols to 0xffffffff
The 0xffffffff shows that we have no m_disabledProtocols
support. The m_disabledProtocols is used only when
POCO_VERSION >= 1.7.0 and in that case, it is initialized
in the constructor to disable some old SSL protocols.

However, if no initial value is given, it reports random
constants to the log which is confusing. Giving it an
initial value makes things clearer.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>