PocoAbstractDao: support transactions
The method session() now returns the session of the current
transaction if any. The dao layer should not create new
transactions as it can be confusing and can cause some
unexpected behaviour when misconfugured. (E.g. when some
SQL statements are executed in context of a different
transaction manager.)

The method session() may ask for a non-transactional
session. The transactional sessions are however preferred.

If a transactional session is asked and there is no transaction
provided in the current transaction manager context, we allocate
a new non-transactional session but drop a warning to log.
When the obsolete non-transactional code is fixed, this warning
should be changed for an exception.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>