Loggable: new abstract class
Loggable provides an easy way how to access a logger
for the current class. The thing is that in C++, it
is not possible to determine the current instance type
in a parent during construction. Thus, calling typeid()
in constructor always returns the current class's typeid.

To avoid this, the Loggable provides method logger that
allows to access the right logger and to ask it on demand.
To avoid logger lookup during runtime, it can be initialized
by setupLogger().

The method logger() and setupLogger() must be const to
allow logging from const methods. Thus, the m_logger is
marked as mutable.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>