AuthService: extract method loginAuthorized
After the authorization of the user is done (some AuthProvider
has authorized the user) we have to open a session for her or
if there is no verified-identity, we have to create one for
that user.

This operation should be executed as a transaction. By extracting
it out of the login() method, we can easily implement the transaction.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>