t: ui: move no-content-length to t/ui/
The no-content-length test is ui-specific as it requires
a HTTP server on the remote site. This bug has not been
discovered as it seems that in all times we were running
both xmlui and ui servers simultaneously.

Currently, automatic integration tests separate running
of both servers and this test fails when running for xmlui.

The test should be run before auth tests is performed, so
change the order a bit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>