LegacyGateway: fix fetching of owner name
The XML-UI protocol expects the name of gateway's owner to be
included in certain responses. We have already supported the
owner field, however, due to lack of testing there was a hidden
bug. The owner's first_name and last_name were not fetched
and the PocoSQLUserDao::parseIfIDNotNull was not called.

This patch fixes the behaviour and adds more testing about this
topic to the SQL query level and to integration tests.

Fixes: a125e5fc5470 ("GatewayXmlHandler: implement handleGet via LegacyGateway")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>