schema: rework cascade-top algorithm
The cascade-top principle is correct however it uses stack to delay
processing of unresolvable elements which leads to infinite loop
in certain cases (never resolves a dependency that is deeper in stack.

This approach works almost the same way but it is simplier and utilizes
a queue instead of stack. The delayed elements are now appended to this
processing queue so it does not freeze in the mentioned case.

Also, handling of self-dependency (loop in the dependency tree) is handled
properly.

In XSLT, the queues must be implemented as pointers/references to a node.
The pointers are numbers stored in a string separated by spaces. To get
the head of a queue, we use substring-before($queue, ' '). Similarly, it
is possible to get all elementes in queue but the head (substring-after).
Appending to queue is performed by concat.

Fixes: c7488871a893 ("schema: generalize drop-tables algorithm")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>