Sanitize: introduce e-mail address sanitization
Implement validation by regular expressions. Validation is performed
for local-part and domain separately. This might be suboptimal but
the code is clear. The validation is based on:

 https://www.w3.org/TR/html/sec-forms.html#valid-e-mail-address
 http://www.regular-expressions.info/email.html

We require the '@' to be present exactly once. We do not permit
any special characters in the domain part. All the address is
treated as case insensitive so the validation is performed
on lower cased address.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>