Sanitize: introduce URI sanitization check
Implement sanitization of URI and fix potential illegal UTF-8
byte sequences. The steps:

1. decode URI
2. check for illegal UTF-8 sequences
3. encode URI
4. construct URI instance that again performs decoding

The result is obtained in step 2, however, we have to prevent
double decoding error so the step 3 is necessary as shown in
attached tests, e.g.:

in: http://www.example.org/%25c0%25ae
1. http://www.example.org/%c0%ae
2. http://www.example.org/%c0%ae
3. http://www.example.org/%25c0%25ae
4. internally in URI http://www.example.org/%c0%ae

However, when calling toString(), the URI would be again encoded
into http://www.example.org/%25c0%25ae.

None of the cases allows to execute the illegal UTF-8 sequence
unintentionally. The only such situation would come when the data
from step 4. are again decoded. This is difficult to happen becase
we have to first export that URI to string and then double decode
the content.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>