schema: unique check (gateway_id, created) in roles_in_gateway
Avoid having any two different identities to have the same created
time in the context of a gateway. We cannot determine the owner for
XML-UI API then.

This has a little drawback, it is necessary to add another role
to a gateway more then 1 second after the first user is associated.
This affects tests.

Tests of failing roles_in_gateway.create are included.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>