Entity: introduce abstract class
A common ancestor of all model objects is helpful as it saves
some code and provides polymorphic access to all of them.
The Entity holds an ID and information whether it has already
been loaded by the DAO layer. Such information is valuable
especially for some security checks and optimizations.

Also, there are overloaded the + operators for easy concatenation
of instances with strings. This is helpful when constructing
log messages.

The EntityLoader abstract class must be inherited by the particular
Dao classes to get access to the setLoaded method which is otherwise
hidden from public.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>