DeviceAccessPolicy: utilize RoleInGatewayDao
The RoleInPlaceDao is not used any more. Every assure method
is now based on gateway and the RoleInGateway.

This simplifies the access level checking because it is not
necessary to query for Gateway (to obtain PlaceID).

The t1001-gateway-register-unregister is temporarily changed
to check for error code 13 (not found). This is not the
finaly state. After the transition to RoleInGateway is done,
the error code will be 10001 again.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>