DefaultAccessPolicy: anybody can unregister from gateway
Anybody who has a role in a gateway can unregister from it.
The lowest role permission that can be obtained in the system
is guest. So, treat guest as the lowest necessary level needed.

If there is no record for the user in database, the fetched
access level would be none. This means that a guest user cannot
unregister from a gateway where it has no role assigned with.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>