RoleXmlHandler: output LegacyRoleInGateway from getall request
The XML-UI API expects to include "name", "surname" and "imgurl".
Also, the "permission" field must contain "owner" for the role
that owns the gateway. This is accomplished by utilizing the
LegacyRoleInGateway and an appropriate serializing method.

As explained in commit:

81ce58405a0a ("query: implement legacy_roles_in_gateway.fetch.by.gateway_id")

the results are longer then expected when a user has more
verified identities for a single identity. This occures in
tests so the tests must be fixed appropriately.

Fixes: 5836aae5b5fe ("RoleXmlHandler: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>