query: implement legacy_roles_in_gateway.fetch.by.gateway_id
Query for legacy roles. This means to gather roles with information
from tables users and verified_identities.

There can be more verified identities assigned to a single identity
and thus the query might return more results then expected for a
single user. Because it cannot be distinguished which verified
identity a role is assigned to, we have to return all of them.
Each verified identity has its own picture URL.

This property might disclose more accounts of a single user to the
audience then expected. To solve this situation, we should enable
to choose which verified identity is the prefered one for an
identity or which data (like picture URL) should be shown to avoid
using verified identities for this purpose (well, it is necessary to
connect an identity to a user).

This issue is left unsolved for now. Currently, only Google is used
as a provider and so this situation might happen only in test cases.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>