SocketServer: provide setter of the server backlog
The backlog can now be set (e.g. from a configuration). Also, the backlog
value is unified for both secure and non-secure server socket. Before,
it was set to the secure server socket only just because there is no
other suitable constructor to pass the Context. Thus, the setup of both
server sockets is now equivalent.

The creation of TCPServer is extracted to its own method to make the
code shorter and better understandable.

Fixes: ad159202f5f0 ("xmlui: utilize SSLServer for SecureServerSocket configuration")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>