DependencyInjector: manage instances wrapped in SharedPtr
This allows to manage the instances in a safer way. Moreover,
it would allow to assign TCPServerConnectionFactory to the TCPServer
without the _disown_ feature.

There is still the same rule:

* create(<name>, false) provides a pointer that is owned by the
  DependencyInjector
* create(<name>, true) provides a pointer that is NOT owned by
  the DependencyInjector and must be deallocated manually

There is however one semantic change, if somebody holds a shared
pointer to some instance at the time when the DependencyInjector
is being deallocated, the instance is not free'd. This means that
the order of deallocation is not preserved in such situation.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>