DependencyInjector: create & find return SharedPtr
The methods create() and find() now return SharedPtr instead
of a raw pointer. This would make the DependencyInjector more
safer according to memory allocation and management.

This change is dangerous because the SharedPtr has implicit
constructor an so there might be some unchecked casts that
lead to memory corruption. Hopefully we don't have any such
leftover.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>