LoopRunner: new class
The LoopRunner is a place from where it is possible to execute
all main application (looping) threads automatically. There might
be a single or multiple LoopRunner instances in the system, each
responsible to start and (finally) stop a set of application threads.

Only stoppable threads can be managable this way. I.e. they must
implement StoppableLoop interface.

The LoopRunner can be started automatically after the dependency
injection is finished for it.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>