StoppableLoop: new class
A StoppableLoop is basically a thread that can be started and stopped.
It manages a thread internally and the thread is not accessible for
the public.

To create a StoppableLoop from StoppableRunnable, the StoppableLoopAdapter
is provided that creates a Poco::Thread for its purpose. Note that it
might leak a Poco::Thread instance in case of misbehaving (join() fails).
This is especially an issue when calling stop() and then start(). However,
it is expected that this occures only when the application is halting.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>