xmlui: utilize LoopRunner to start everything
There is no longer any important piece of network code in the main
routine. All main application threads are handled by the LoopRunner.

Note that the port configuration has been moved to the conf/server.ini
and we allow to override it from command line using a kind of hack.
This is to be fixed later by unifying startup even more.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>