di: prefix all BEEEON_OBJECT names by namespace
The first argument of BEEEON_OBJECT is an identificator used by
linker during compilation. Thus, it should be unique and it must
not interfere with other projects. Use prefix derived from the
namespace.

Unfortunately, it seems that there is no way how to generated this
prefix from the real class prefix automatically using C macros.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>